clinical-parse: move trim_whitespace function
Refactor the `trim_whitespace` function and its helper function to a
shared module for reuse by future commands.
